chore(federation): split operation contains functions
