+++ apollo-federation/src/operation/tests/mod.rs
@@ -3,8 +3,6 @@ use apollo_compiler::ExecutableDocument;
 use indexmap::IndexSet;
 
 use super::normalize_operation;
-use super::Containment;
-use super::ContainmentOptions;
 use super::Name;
 use super::NamedFragments;
 use super::Operation;
@@ -1611,193 +1609,6 @@ type T {
             assert_eq!(actual, expected);
         }
     }
-}
-
-fn containment_custom(left: &str, right: &str, ignore_missing_typename: bool) -> Containment {
-    let schema = apollo_compiler::Schema::parse_and_validate(
-        r#"
-        directive @defer(label: String, if: Boolean! = true) on FRAGMENT_SPREAD | INLINE_FRAGMENT
-
-        interface Intf {
-            intfField: Int
-        }
-        type HasA implements Intf {
-            a: Boolean
-            intfField: Int
-        }
-        type Nested {
-            a: Int
-            b: Int
-            c: Int
-        }
-        input Input {
-            recur: Input
-            f: Boolean
-            g: Boolean
-            h: Boolean
-        }
-        type Query {
-            a: Int
-            b: Int
-            c: Int
-            object: Nested
-            intf: Intf
-            arg(a: Int, b: Int, c: Int, d: Input): Int
-        }
-        "#,
-        "schema.graphql",
-    )
-    .unwrap();
-    let schema = ValidFederationSchema::new(schema).unwrap();
-    let left = Operation::parse(schema.clone(), left, "left.graphql", None).unwrap();
-    let right = Operation::parse(schema.clone(), right, "right.graphql", None).unwrap();
-
-    left.selection_set.containment(
-        &right.selection_set,
-        ContainmentOptions {
-            ignore_missing_typename,
-        },
-    )
-}
-
-fn containment(left: &str, right: &str) -> Containment {
-    containment_custom(left, right, false)
-}
-
-#[test]
-fn selection_set_contains() {
-    assert_eq!(containment("{ a }", "{ a }"), Containment::Equal);
-    assert_eq!(containment("{ a b }", "{ b a }"), Containment::Equal);
-    assert_eq!(
-        containment("{ arg(a: 1) }", "{ arg(a: 2) }"),
-        Containment::NotContained
-    );
-    assert_eq!(
-        containment("{ arg(a: 1) }", "{ arg(b: 1) }"),
-        Containment::NotContained
-    );
-    assert_eq!(
-        containment("{ arg(a: 1) }", "{ arg(a: 1) }"),
-        Containment::Equal
-    );
-    assert_eq!(
-        containment("{ arg(a: 1, b: 1) }", "{ arg(b: 1 a: 1) }"),
-        Containment::Equal
-    );
-    assert_eq!(
-        containment("{ arg(a: 1) }", "{ arg(a: 1) }"),
-        Containment::Equal
-    );
-    assert_eq!(
-        containment(
-            "{ arg(d: { f: true, g: true }) }",
-            "{ arg(d: { f: true }) }"
-        ),
-        Containment::NotContained
-    );
-    assert_eq!(
-        containment(
-            "{ arg(d: { recur: { f: true } g: true h: false }) }",
-            "{ arg(d: { h: false recur: {f: true} g: true }) }"
-        ),
-        Containment::Equal
-    );
-    assert_eq!(
-        containment("{ arg @skip(if: true) }", "{ arg @skip(if: true) }"),
-        Containment::Equal
-    );
-    assert_eq!(
-        containment("{ arg @skip(if: true) }", "{ arg @skip(if: false) }"),
-        Containment::NotContained
-    );
-    assert_eq!(
-        containment("{ ... @defer { arg } }", "{ ... @defer { arg } }"),
-        Containment::NotContained,
-        "@defer selections never contain each other"
-    );
-    assert_eq!(
-        containment("{ a b c }", "{ b a }"),
-        Containment::StrictlyContained
-    );
-    assert_eq!(
-        containment("{ a b }", "{ b c a }"),
-        Containment::NotContained
-    );
-    assert_eq!(containment("{ a }", "{ b }"), Containment::NotContained);
-    assert_eq!(
-        containment("{ object { a } }", "{ object { b a } }"),
-        Containment::NotContained
-    );
-
-    assert_eq!(
-        containment("{ ... { a } }", "{ ... { a } }"),
-        Containment::Equal
-    );
-    assert_eq!(
-        containment(
-            "{ intf { ... on HasA { a } } }",
-            "{ intf { ... on HasA { a } } }",
-        ),
-        Containment::Equal
-    );
-    // These select the same things, but containment also counts fragment namedness
-    assert_eq!(
-        containment(
-            "{ intf { ... on HasA { a } } }",
-            "{ intf { ...named } } fragment named on HasA { a }",
-        ),
-        Containment::NotContained
-    );
-    assert_eq!(
-        containment(
-            "{ intf { ...named } } fragment named on HasA { a intfField }",
-            "{ intf { ...named } } fragment named on HasA { a }",
-        ),
-        Containment::StrictlyContained
-    );
-    assert_eq!(
-        containment(
-            "{ intf { ...named } } fragment named on HasA { a }",
-            "{ intf { ...named } } fragment named on HasA { a intfField }",
-        ),
-        Containment::NotContained
-    );
-}
-
-#[test]
-fn selection_set_contains_missing_typename() {
-    assert_eq!(
-        containment_custom("{ a }", "{ a __typename }", true),
-        Containment::Equal
-    );
-    assert_eq!(
-        containment_custom("{ a b }", "{ b a __typename }", true),
-        Containment::Equal
-    );
-    assert_eq!(
-        containment_custom("{ a b }", "{ b __typename }", true),
-        Containment::StrictlyContained
-    );
-    assert_eq!(
-        containment_custom("{ object { a b } }", "{ object { b __typename } }", true),
-        Containment::StrictlyContained
-    );
-    assert_eq!(
-        containment_custom(
-            "{ intf { intfField __typename } }",
-            "{ intf { intfField } }",
-            true
-        ),
-        Containment::StrictlyContained,
-    );
-    assert_eq!(
-        containment_custom(
-            "{ intf { intfField __typename } }",
-            "{ intf { intfField __typename } }",
-            true
-        ),
-        Containment::Equal,
-    );
 }
 
 /// This regression-tests an assumption from
